reduce slippage in tickless sleep
Prior to this change, every execution of vPortSuppressTicksAndSleep( ) stopped the systick.  
After this change, vPortSuppressTicksAndSleep( ) stops the systick only after calling eTaskConfirmSleepModeStatus( ), and only if that function confirms the sleep status (return code *not* eAbortSleep).  This change reduces drift/slippage in kernel time because some of the calls to vPortSuppressTicksAndSleep( ) will now have zero drift.