vTaskStepTick( ) assertion refinement
vPortSuppressTicksAndSleep( ) should not jump the tick count all the way to xNextTaskUnblockTime.  Reaching xNextTaskUnblockTime should instead occur in xTaskIncrementTick( ).  In the case where vPortSuppressTicksAndSleep( ) sleeps the entire expected time, the final tick to reach xNextTaskUnblockTime will be in xPendedTicks, where it can be processed normally in xTaskIncrementTick( ), when the idle task resumes the scheduler after the sleep.