Override SB trigger level for small reads
Consider that some callers of xStreamBufferReceive() use parameter
xBufferLengthBytes to limit the number of bytes received for efficient
parsing of the stream.

Prior to this commit, if a task calls xStreamBufferReceive() to receive
fewer bytes than the trigger level, and if xStreamBufferReceive()
decides to block, then the task remains blocked even after the desired
number of bytes arrive in the stream buffer.  When the receive operation
finally times out, xStreamBufferReceive() returns the desired number of
bytes to fill the caller's buffer.  To avoid the timeout, the task must
call xStreamBufferSetTriggerLevel() before xStreamBufferReceive().

After this commit, xStreamBufferReceive() automatically and temporarily
overrides the trigger level to ensure blocking ends when the desired
number of bytes arrives into the stream buffer.

This commit allows applications to avoid comparing the desired number of
bytes to the trigger level before every stream read.  The code in this
commit does that same comparison but only if it will block.