Check trigger level before unblocking SB sender
The stream buffer implementation waits for the contents of the buffer to
reach a trigger level before unblocking the stream buffer receiver.
This commit duplicates that behavior for unblocking the stream buffer
sender.

The trigger level should apply to *both* a blocked receiver and a
blocked sender.  The purpose of the trigger level is to avoid starving
lower-priority tasks unnecessarily.  For example, a receiver task may be
blocked waiting for additional streaming bytes, but it processes groups
of bytes more efficiently than single bytes.  The same reasoning applies
to a sender task that sends groups of bytes more efficiently than single
bytes.

Backward Compatibility

Applications that set the trigger level to 2 or higher may be affected.
However, of those applications, only those with sending tasks that block
are actually affected.  With this commit, those blocked sending tasks
may unblock later than before if the receiving task reads an amount of
data smaller than the trigger level.  Note however that the conditions
at the time the sender is "finally" unblocked match the conditions had
the receiver simply read a larger amount of data in the first place.
The sender almost certainly handles that case in any application code
affected by this change.  Thus this commit can be considered backward
compatible, and the next upgrade guide and API docs should indicate the
new behavior.