Add option to defer start/reset timer from ISR
Prior to this commit, functions xTimerStartFromISR() and
xTimerResetFromISR() may not behave as expected when called from an ISR
that interrupts a tickless idle period.  The expected start time (or
reset time) of the timer is the time at which the ISR calls the timer
function, but prior to this commit the actual start time (or reset time)
was the time at the beginning of the tickless idle period.

The root cause is a stale value being returned by
xTaskGetTickCountFromISR().  The value is stale because the ISR executed
before the tickless logic corrected the system tick count.  This commit
adds a configuration option allowing the timer task to replace the stale
tick count with the current tick count.

In configurations with tickless idle disabled, the new option defaults
to disabled, for backward compatibility.  With tickless idle enabled,
the new option defaults to enabled, which represents a change in
behavior from before this commit.  The changed behavior is closer to
expected behavior for tickless implementations that allow the ISR to
execute prior to correcting the system tick count.

Tickless-idle implementations that correct the system tick count prior
to allowing the ISR to execute may explicitly disable the new option
as those systems have no need for this option.